fix: transaction activity page
This commit fixes the orders displayed on the transaction activity page.

I added a check if the orders array is empty and if it is, I display a message indicating that there are no transaction activity to display if its not i display the activities

I also added optional chain to the orders array to prevent errors when trying to access properties of undefined objects.

The commit also incudes some minor optional chaining in the grabbed product page to prevent errors when trying to access properties of undefined objects.