Fix/transaction activity page (#94)
* Implemented grabbed product flow uis, worked on some corrections

* corrected frontend link

* Escrow integration, implemented requested changes

* Started Logistics, worked on  Pickup & Delivery pages, added address to payload

* conflict fixed

* Delivery flow changed, Fixed post ad buttons

* Logistic orders page

* logistics dashboard

* Logistics services yanked off, work ongoing on integrating escrow and quote feature for delivery

* Implemented safety guides page, changed the coin amount displayed for Basic feature

* conflict merged

* Checkout implemented

* Removed console.log

* fix bugs

* fix: transaction activity page

This commit fixes the orders displayed on the transaction activity page.

I added a check if the orders array is empty and if it is, I display a message indicating that there are no transaction activity to display if its not i display the activities

I also added optional chain to the orders array to prevent errors when trying to access properties of undefined objects.

The commit also incudes some minor optional chaining in the grabbed product page to prevent errors when trying to access properties of undefined objects.