fix(logout):this fixes the notification of logout
This PR brings in a clean up for logout, but will have a follow up issue
to rework the logout and axios implementation. The issue was that the
socket was trying to connect to backend and failed.
We don't need that error. A reload fixes it.

This PR will be referenced, when the follow up PR is made.
Closes: https://github.com/afficode/frontend/issues/141

DCO 1.1 Signed-off-by Samuel Chika <samuelemyrs@gmail.com>