fix(sec and login): security and login (#161)
This PR addresses issue: https://github.com/afficode/frontend/issues/143

The Avatar icon is now implement with the user credentials
The Auto Dealer is now scrapped along side the other 2 options of save login info and get notification of unrecognized devices.

This is because we don't support this yet and we can revisit it when we have the infrastructure for it.

Closes: https://github.com/afficode/frontend/issues/143

DCO 1.1 Signed-off-by Samuel Chika <samuelemyrs@gmail.com>