feat(close_ad): close advert from the Advert page in the profile
This PR address the close button in the Advert page.

The Advert page close button for each advert displayed wasn't working. But this PR introduce the logic to it.
What this PR addresses is:
- make the close button functional.
- The filter buttons for each advert condition is now functional (All, Active, Blocked, Closed).
- The Blocked Ad introduced 2 option which is: a. when the Ad is not yet paid for, and b. when paid but blocked.
- Users can now make payment from their blocked Ad (regardless if its from the blocked filter or the all filter).
- Reduce the payload size from the backend to capture only required field.

This will enhance speed to fetch payload.

Every action carried out will be seen in effect.

This has been tested in the backend and it works fine.

Closes issue: https://github.com/afficode/frontend/issues/149

DCO 1.1 Signed-off-by Samuel Chika <samuelemyrs@gmail.com>