fix(url_state_search):
This PR address the LGA and State url link under the title when viewing an Ad.

The problem is: when you click on the link, the URL shows the PAGE NOT FOUND. This is because the URL was wrong.

The URL is now fixed and everything works.

This closes issue: https://github.com/afficode/frontend/issues/144

DCO 1.1 Signed-off-by Samuel Chika <samuelemyrs@gmail.com>