fix(update user): update phone number in browser
This PR addresses the issue of verify phone number.
Previously, users after verifying phone number, have to re-auth before seeing effect, there by showing a UI of success and still having verify phone button.

This PR updates the user phone number in the user payload when a successful verification is done.

Users don't need to re-auth.

Also, the phone number payload will not make it to the backend if it did not change.  If its changed, the user payload is updated and the user have to re-verify the number.

DCO 1.1 Signed-off-by Samuel Chika <samuelemyrs@gmail.com>