feat(escrow): finalizing the escrow service feature.
This PR introduces a finalization of the escrow service to align with
business objectives and BE implementation.

This PR modifies the TrransactionActivity page UI to look more modern
and appeasing to the user, by introducing some icons in the table to
inform user about the state of their transaction.

THis PR also changed the pill display of badges in table values to look
like a modern dashboard.

We also introduces the Refund Processing option, which ensures that a
user cannot proceed to releasing fund when they have already book in for
a refund. We display a loader icon on the table and also a text to the
user acknowledging we are processing the Refund and a follow up
information.

We also added some minor fixes where axios keep littering the browser
console wth 404 error, which is bound to happen because this is a check
for order availablilty on an AD, which most times, it will be 404.
THis is fixed by the backend returning 204 instead of 404.

DCO Signed Off: Samuel Ezeja <samuelemyrs@gmail.com>