chore(cleanup): cleanup and policy implementation
A new policy is introduced to change the size of images to be uploaded from 10 to 5.

With this PR, the image settings is now standardized in the boonfuInfo.js with specific values which if policy changes again it will be easy to make changes to the code.

This also remove all the unnecessary navbar options which are not in use.

DCO Signed Off: Samuel Ezeja <samuelemyrs@gmail.com>